Needs Cake Screen Finished 99%

+++ src/game/level_update.h
@@ -180,7 +180,7 @@ s32 lvl_init_or_update(                  s16 initOrUpdate, UNUSED s32 levelNum);
 s32 lvl_init_from_save_file(      UNUSED s16 initOrUpdate,        s32 levelNum);
 s32 lvl_set_current_level(        UNUSED s16 initOrUpdate,        s32 levelNum);
 s32 lvl_play_the_end_screen_sound(UNUSED s16 initOrUpdate, UNUSED s32 levelNum);
-s8 print_winnings_text();
+s16 print_winnings_text();
 void basic_update(void);
 
 #endif // LEVEL_UPDATE_H
